# evechem_api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "Evechem API"
schemes:
- "https"
paths:
...
  /towers/:
    get:
      tags:
      - "Starbase"
      description: "Gets summary information for every tower the request is authenticated\
        \ to see.  This includes `tower_id` number, `type_id`, fuel status, as well\
        \ as what the tower is producing at the endpoints of its reaction/mining chains.\n"
      operationId: "towers_get"
      produces:
      - "application/json"
      parameters:
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            description: "Array of tower id numbers the authenticated user can interact\
              \ with.\n"
            items:
              type: "integer"
        401:
          description: "Authentication Required"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    post:
      tags:
      - "Starbase"
      description: "Add new tower.\n"
      operationId: "towers_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "tower"
        description: "New tower information in json format."
        required: true
        schema:
          $ref: "#/definitions/new_tower"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
  /towers/{tower_id}/:
    get:
      tags:
      - "Starbase"
      description: "Gets details for a specific tower with `type_id`.\n"
      operationId: "towers_tower_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Type of the item that needs fetched."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/tower_details"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    delete:
      tags:
      - "Starbase"
      description: "Delete a specific tower and all its equipment/processes.\n"
      operationId: "towers_tower_id_delete"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Id number of the tower to be deleted."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    patch:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_patch"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Id number of the tower to be deleted."
        required: true
        type: "integer"
      - in: "body"
        name: "tower_update"
        description: "Details to update tower with."
        required: false
        schema:
          $ref: "#/definitions/tower_details"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/tower_details"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
  /towers/{tower_id}/processes/:
    get:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_get"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/process"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    post:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/process"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
  /towers/{tower_id}/processes/{process_id}/:
    get:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/process"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    delete:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_delete"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Authentication Required"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
  /towers/{tower_id}/processes/{process_id}/equipment/:
    get:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_equipment_get"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/equipment"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    post:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_equipment_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - in: "body"
        name: "new_equipment"
        required: true
        schema:
          $ref: "#/definitions/new_equipment"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/equipment"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/error"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
  /towers/{tower_id}/processes/{process_id}/equipment/{equipment_id}/:
    get:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_equipment_equipment_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - name: "equipment_id"
        in: "path"
        description: "Id of equipment that needs fetched."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/equipment"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    delete:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_equipment_equipment_id_delete"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - name: "equipment_id"
        in: "path"
        description: "Id of equipment that needs fetched."
        required: true
        type: "integer"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
    patch:
      tags:
      - "Starbase"
      operationId: "towers_tower_id_processes_process_id_equipment_equipment_id_patch"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "tower_id"
        in: "path"
        description: "Tower Id number to look under."
        required: true
        type: "integer"
      - name: "process_id"
        in: "path"
        description: "Process Id number to look under."
        required: true
        type: "integer"
      - name: "equipment_id"
        in: "path"
        description: "Id of equipment that needs fetched."
        required: true
        type: "integer"
      - in: "body"
        name: "equipment"
        required: true
        schema:
          $ref: "#/definitions/equipment_update"
      - name: "api_key"
        in: "query"
        description: "Operation Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/equipment"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/error"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.starbase_controller"
definitions:
  equipment:
    type: "object"
    required:
    - "id"
    - "inputs"
    - "outputs"
    - "resource"
    properties:
      contains:
        type: "integer"
        description: "Quantity of material stored when last updated.\n"
      id:
        type: "integer"
        description: "Unique id number for this equipment.\n"
      inputs:
        type: "array"
        description: "Array of input links.\n"
        items:
          $ref: "#/definitions/link"
      last_updated:
        type: "integer"
      outputs:
        type: "array"
        description: "Array of output links.\n"
        items:
          $ref: "#/definitions/link"
      resource:
        type: "integer"
        description: "type_id of material/reaction being stored/mined/processed in\
          \ this equipment.\n"
    title: "equipment"
...
  equipment_update:
    type: "object"
    required:
    - "last_updated"
    properties:
      contains:
        type: "integer"
      inputs:
        type: "array"
        description: "Array of input links.\n"
        items:
          $ref: "#/definitions/link"
      last_updated:
        type: "integer"
      online:
        type: "boolean"
      outputs:
        type: "array"
        description: "Array of output links.\n"
        items:
          $ref: "#/definitions/link"
      resource:
        type: "integer"
    title: "equipment_update"
  error:
    type: "object"
    properties:
      message:
        type: "string"
    title: "error"
...
  link:
    type: "object"
    required:
    - "equipment"
    - "material"
    properties:
      equipment:
        type: "integer"
        description: "Equipment that is the source/target of this link.\n"
      input:
        type: "boolean"
        description: "Optional boolean designating if this in an input or output link.\n"
      material:
        type: "integer"
        description: "type_id of the linked material.\n"
    title: "equipment_link"
  location:
    type: "object"
    required:
    - "moon"
    - "planet"
    properties:
      moon:
        type: "integer"
      planet:
        type: "integer"
      system:
        $ref: "#/definitions/location_system"
  location_system:
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
...
  new_equipment:
    type: "object"
    required:
    - "type"
    properties:
      contains:
        type: "integer"
      name:
        type: "string"
      online:
        type: "boolean"
      resource:
        type: "integer"
      type:
        type: "integer"
    title: "new_equipment"
  new_tower:
    type: "object"
    required:
    - "location"
    - "online"
    - "type"
    properties:
      cycles_at:
        type: "integer"
        description: "Minute (0-59) in which this tower processes all fuel and silo/reaction\
          \ cycles.\n"
        minimum: 0
        maximum: 59
      fuel_count:
        type: "integer"
      fuel_last_udpate:
        type: "integer"
      location:
        $ref: "#/definitions/location"
      online:
        type: "boolean"
        description: "Online/offline status boolean.\n"
      owner:
        type: "integer"
      type:
        type: "integer"
    title: "new_tower"
  process:
    type: "object"
    properties:
      equipment:
        type: "array"
        items:
          $ref: "#/definitions/equipment"
      final_outputs:
        type: "array"
        description: "Final products' material ids for this process tree.\n"
        items:
          type: "integer"
      id:
        type: "number"
        description: "Unique process id number.\n"
    title: "process"
...
  tower_details:
    type: "object"
    properties:
      cycles_at:
        type: "integer"
        description: "Minute (0-59) in which this tower processes all fuel and silo/reaction\
          \ cycles.\n"
        minimum: 0
        maximum: 59
      fuel_count:
        type: "integer"
      fuel_last_update:
        type: "integer"
      id:
        type: "integer"
      location:
        $ref: "#/definitions/location"
      name:
        type: "string"
      online:
        type: "boolean"
        description: "Online/offline status boolean.\n"
      processes:
        type: "array"
        description: "Process Ids that this tower has set up.\n"
        items:
          type: "integer"
      sov:
        type: "boolean"
        description: "Whether or not the tower is anchored with a sovereignty fuel\
          \ bonus active.\n"
      type:
        type: "integer"
    title: "starbase_tower"
...